+++ resources/js/Layouts/FooterLayout.vue
